+++ src/optioncodes.ts
@@ -1,6 +1,4 @@
-'use strict'
-
-exports.toString = function (type) {
+export function toString (type: number) {
   switch (type) {
     // list at
     // https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-11
@@ -25,7 +23,7 @@ exports.toString = function (type) {
   return `OPTION_${type}`
 }
 
-exports.toCode = function (name) {
+export function toCode (name: string | number) {
   if (typeof name === 'number') {
     return name
   }
